src/User/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
PyJWT==2.6.0
Werkzeug==2.2.2
Flask-CORS==5.0.0
PyMySQL==1.0.3
Flask-PyMongo==2.3.0
pika
requests